Fix GVFS client can't find `HDFSFileSystemProvider` problem
